Moved Instrumentation tests from unit test file to Instrumentation test file.
